Ged rid of BidPrivider and convert it to custom hooks
