+++ src/components/Tables/DomainTable/components/NFTCardActions.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect, useRef } from 'react';
 
 // Library Imports
 import { Domain } from 'lib/types';
-import { useBidProvider } from 'lib/providers/BidProvider';
+import { useBidProvider } from 'lib/hooks/useBidProvider';
 
 // Component Imports
 import { Spinner } from 'components';
