+++ src/containers/flows/MakeABid/MakeABid.tsx
@@ -6,7 +6,7 @@ import { useWeb3React } from '@web3-react/core'; // Wallet data
 import { Web3Provider } from '@ethersproject/providers/lib/web3-provider'; // Wallet data
 //- Library Imports
 import { Domain, Metadata, Bid, Maybe } from 'lib/types';
-import { useBidProvider } from 'lib/providers/BidProvider';
+import { useBidProvider } from 'lib/hooks/useBidProvider';
 import { getMetadata } from 'lib/metadata';
 import { toFiat } from 'lib/currency';
 import useCurrency from 'lib/hooks/useCurrency';
