+++ src/lib/hooks/useZAuctionBaseApiUri.ts
@@ -1,24 +1,11 @@
-import React, { useState } from 'react';
-
 //the hook that returns the useZauctionAPI with the api endpoint that match with the selected chain
+// Infact it is not HOOK! Just Utility!
 
-export function useZAuctionBaseApiUri(chain: number): string | undefined {
-	const [api, setApi] = useState<string | undefined>();
+export const ZAUCTION_BASE_URL: { [key: number]: string } = {
+	1: 'https://mainnet.zauction.api.zero.tech/api',
+	42: 'https://zauction-kovan-api.herokuapp.com/api',
+};
 
-	React.useEffect(() => {
-		function getApiEndpoint() {
-			switch (chain) {
-				case 1:
-					return 'https://mainnet.zauction.api.zero.tech/api';
-				case 42:
-					return 'https://zauction-kovan-api.herokuapp.com/api';
-				default:
-					return 'Unsupported Chain';
-			}
-		}
-
-		setApi(getApiEndpoint());
-	}, [chain]);
-
-	return api;
-}
+export const useZAuctionBaseApiUri = (chain: number): string | undefined => {
+	return ZAUCTION_BASE_URL[chain];
+};
